feat(includers/generic): include markdown content
introduce new includer - generic includer

includes folder with markdown content

soft deprication of the sourcedocs includer in favor of the generic one
