+++ src/services/includers/batteries/index.ts
@@ -1,2 +1,3 @@
+export * as generic from './generic';
 export * as sourcedocs from './sourcedocs';
 export * as openapi from './openapi';
